+++ Import/AeteObject.m
@@ -10,7 +10,7 @@
 
 @implementation SdefObject (AeteResource)
 
-- (UInt32)parseData:(Byte *)bytes {
+- (NSUInteger)parseData:(Byte *)bytes {
   ShadowTrace();
   [NSException raise:NSInternalInconsistencyException format:@"Method must be implemented by subclasses"];
   return 0;
